Reorganise telegraph service
