+++ application/services/telegraph/src/user/controllers/user/user.controller.ts
@@ -2,7 +2,7 @@ import {EventPattern, Payload} from '@looport/nest-microservices'
 import {Controller} from '@nestjs/common'
 
 import {UserRepository} from '@/storage/repositories/user/user.repository'
-import {UserDto} from '@/user/dtos/user/user.dto'
+import {UserDto} from '@/user/dto/user/user.dto'
 
 @Controller()
 export class UserController {
