+++ application/services/telegraph/src/user/user.module.ts
@@ -2,7 +2,7 @@ import {Module} from '@nestjs/common'
 
 import {StorageModule} from '@/storage/storage.module'
 
-import {UserController} from './user.controller'
+import {UserController} from './controllers/user/user.controller'
 
 @Module({
   controllers: [UserController],
